enhance missing job detection
Pull available jobs from jenkins api instead of manually looking up every
url to see if it exists

+++ purger.py
@@ -1,3 +1,4 @@
+import time
 import requests
 from mc_bin_client import MemcachedClient as McdClient
 
@@ -5,17 +6,43 @@ BUCKETS=['server', 'mobile', 'sdk']
 MC_HOST="127.0.0.1"
 MC_PORT=11210
 VIEW_API="http://172.23.121.132:8092/"
+JENKINS_URLS=["http://qa.sc.couchbase.com/",
+             "http://qa.hq.northscale.net/",
+             "http://ci.sc.couchbase.com/",
+             "http://sdkbuilds.couchbase.com/",
+             "http://mobile.jenkins.couchbase.com/",
+             "http://sdkbuilds.couchbase.com/view/LCB/job/feature/",
+             "http://sdkbuilds.couchbase.com/view/LCB/job/situational-lcb/",
+             "http://sdkbuilds.couchbase.com/view/JAVA/job/situational-java/",
+             "http://sdkbuilds.couchbase.com/view/.NET/",
+             "http://qa.sc.couchbase.com/view/extended/"]
+client = McdClient(MC_HOST, MC_PORT)
 
-for bucket in BUCKETS:
 
-    client = McdClient(MC_HOST, MC_PORT)
+def queryJenkinsJobs():
+    _JOBS = []
+
+    for url in JENKINS_URLS:
+        url=url+"api/json"
+        r = requests.get(url)
+        if r.status_code == 200:
+            data = r.json()
+            jobs = data.get("jobs")
+            job_names = [str(n['name']) for n in jobs]
+            _JOBS.extend(job_names)
+        else:
+            "ERROR Jenkins url: (%s, %s)" % (r.status_code, url)
+
+    return _JOBS
+
+def purge(bucket, known_jobs):
+
+
     client.sasl_auth_plain(bucket, "")
-
     DDOC='/'.join([VIEW_API,bucket,'_design',bucket,'_view'])
     PLATFORMS_Q=DDOC+"/all_platforms?stale=false&group_level=1&inclusive_end=true&reduce=true"
     COMPONENTS_Q=DDOC+"/all_components?full_set=true&group_level=1&inclusive_end=true&reduce=true&stale=false"
     BUILDS_Q = DDOC+"/data_by_build?full_set=true&group_level=1&inclusive_end=true&reduce=true&stale=false"
-
     ALL_PLATFORMS = []
     ALL_COMPONENTS = []
 
@@ -59,13 +86,14 @@ for bucket in BUCKETS:
             cnt = val[5]
             bid = val[6]
 
-            # try to get job url
-            url = val[3]
-            r = requests.get(url)
-            if r.status_code == 404:
-                print "****MISSING*** %s_%s: %s:%s:%s (%s,%s)" % (build, _id, os, comp, name, val[5], bid)
-                client.delete(_id, vbucket=0)
-                continue
+            # if job is unkown try to manually get url
+            if name not in known_jobs:
+                url = val[3]
+                r = requests.get(url)
+                if r.status_code == 404:
+                    print "****MISSING*** %s_%s: %s:%s:%s (%s,%s)" % (build, _id, os, comp, name, val[5], bid)
+                    client.delete(_id, vbucket=0)
+                    continue
 
             if os in JOBS:
                 if comp in JOBS[os]:
@@ -98,3 +126,14 @@ for bucket in BUCKETS:
                 JOBS[os] = {}
                 JOBS[os][comp] = [(name, bid, _id)]
 
+if __name__ == "__main__":
+    while True:
+        try:
+            known_jobs = queryJenkinsJobs()
+            for bucket in BUCKETS:
+                purge(bucket, known_jobs)
+        except Exception as ex:
+            print ex
+            pass
+
+        time.sleep(500)
